Refactor about_section import and layout
